MacUpdater: check os version
