+++ Source/Core/Common/build_info.txt.in
@@ -2,7 +2,6 @@
 // Updater will fail the update if the user does not meet this requirement.
 OSMinimumVersionWin10=10.0.15063.0
 OSMinimumVersionWin11=10.0.22000.0
-OSMinimumVersionMacOS=10.14
 
 // This is the runtime which was compiled against - providing a way for Updater to detect if update
 // is needed before executing this binary. Note that, annoyingly, the version in environment
